requirements.txt
fastapi
uvicorn
qrcode[pil]
python-multipart
email-validator
pillow
jinja2